// TaxiServiceSPA/src/interfaces/models/driver.ts
export interface Driver {
    driverId: number;
    departmentId: number;
    carId: number;
    lastName: string;
    firstName: string;
    middleName: string;
    dateOfBirth: string;
    region: string;
    city: string;
    street: string;
    building: string;
    flat: number;
    beginning: string;
    ending: string;
    salary: number;
    available: boolean;
}